+++ nflow-jetty/src/test/java/io/nflow/jetty/mapper/IllegalArgumentExceptionMapperTest.java
@@ -1,4 +1,4 @@
-package io.nflow.rest.mapper;
+package io.nflow.jetty.mapper;
 
 import static javax.ws.rs.core.Response.Status.BAD_REQUEST;
 import static org.hamcrest.CoreMatchers.is;
@@ -8,16 +8,15 @@ import javax.ws.rs.core.Response;
 
 import org.junit.jupiter.api.Test;
 
-import io.nflow.engine.workflow.executor.StateVariableValueTooLongException;
-import io.nflow.rest.mapper.StateVariableValueTooLongExceptionMapper;
+import io.nflow.jetty.mapper.IllegalArgumentExceptionMapper;
 import io.nflow.rest.v1.msg.ErrorResponse;
 
-public class StateVariableValueTooLongExceptionMapperTest {
-  StateVariableValueTooLongExceptionMapper mapper = new StateVariableValueTooLongExceptionMapper();
+public class IllegalArgumentExceptionMapperTest {
+  IllegalArgumentExceptionMapper mapper = new IllegalArgumentExceptionMapper();
 
   @Test
   public void exceptionIsMappedToBadRequest() {
-    try (Response response = mapper.toResponse(new StateVariableValueTooLongException("error"))) {
+    try (Response response = mapper.toResponse(new IllegalArgumentException("error"))) {
       assertThat(response.getStatus(), is(BAD_REQUEST.getStatusCode()));
       ErrorResponse error = (ErrorResponse) response.getEntity();
       assertThat(error.error, is("error"));
